First commit - Template10 boilerplate + Preliminary FB Auth
This first commit introduces the groundwork for the project. This means the inclusion of "Template10" as the framework of choice. Given that Universal Apps and XAML is new to me. I figured a pre-existing wrapper that abstracts a little bit of the nitty gritty to be useful.

I've implemented a test LoginPage that calls FB to retrieve an authorization token. This code will need to be moved to the model in the future.