App Open Logic and Match View Implementation
Implemented /user/ping (change location) as well as the logic to determine if a user has to login again or re-use existing XAUthToken.

Preliminary work has been done on the SuperficialPage to bind getMatches() to a ListView. Going forward this needs to be much better but is a proof of concept that it works :)