First version of SuperficialPage
The superficial page is where matching/passing is performed. This is the
first commit in terms of actually retrieving the first match from the
API. Most of the work in this commit was on UI binding and retrieving
the first info. More work needs to be done on actual processing of the
matches (pass/like/superlike/retrieve next match)