SuperficialPage: See who you will be reviewing next before deciding on the current.
Tinder has a neat little feature when swiping left and right. It lets you see who is going to come next. To impleemnt this on Tindows we make it such that a "NextInLine" property is always available and up to date with the next match in the queue.

From there, same display logic as before.